scripts/fetch_tasks.py
"""
Task Dataset Scraper

Downloads programming tasks from public datasets:
- HumanEval
- MBPP (Mostly Basic Python Problems)
- CodeNet (selected problems)
"""

import json
import requests
from pathlib import Path
from typing import List, Dict, Any
import time

from src.utils.logger import setup_logger

logger = setup_logger(__name__)


class HumanEvalScraper:
    """Scraper for HumanEval dataset."""
    
    HUMANEVAL_URL = "https://raw.githubusercontent.com/openai/human-eval/master/data/HumanEval.jsonl.gz"
    
    def __init__(self):
        self.tasks = []
    
    def fetch_tasks(self, num_tasks: int = 10) -> List[Dict[str, Any]]:
        """
        Fetch tasks from HumanEval dataset.
        
        Args:
            num_tasks: Number of tasks to fetch
            
        Returns:
            List of task dictionaries
        """
        logger.info(f"Fetching {num_tasks} tasks from HumanEval")
        
        # TODO: Implement HumanEval fetching
        # For now, return mock tasks
        tasks = []
        
        logger.info(f"Fetched {len(tasks)} HumanEval tasks")
        return tasks


class MBPPScraper:
    """Scraper for MBPP dataset."""
    
    MBPP_URL = "https://raw.githubusercontent.com/google-research/google-research/master/mbpp/mbpp.jsonl"
    
    def __init__(self):
        self.tasks = []
    
    def fetch_tasks(self, num_tasks: int = 10) -> List[Dict[str, Any]]:
        """
        Fetch tasks from MBPP dataset.
        
        Args:
            num_tasks: Number of tasks to fetch
            
        Returns:
            List of task dictionaries
        """
        logger.info(f"Fetching {num_tasks} tasks from MBPP")
        
        # TODO: Implement MBPP fetching
        tasks = []
        
        logger.info(f"Fetched {len(tasks)} MBPP tasks")
        return tasks


class CodeNetScraper:
    """Scraper for CodeNet dataset."""
    
    def __init__(self):
        self.tasks = []
    
    def fetch_tasks(self, num_tasks: int = 5) -> List[Dict[str, Any]]:
        """
        Fetch tasks from CodeNet dataset.
        
        Args:
            num_tasks: Number of tasks to fetch
            
        Returns:
            List of task dictionaries
        """
        logger.info(f"Fetching {num_tasks} tasks from CodeNet")
        
        # TODO: Implement CodeNet fetching
        # CodeNet is large; we'll need to select specific problems
        tasks = []
        
        logger.info(f"Fetched {len(tasks)} CodeNet tasks")
        return tasks


class TaskDatasetBuilder:
    """Builds task dataset from multiple sources."""
    
    def __init__(self, output_dir: Path = Path("tasks")):
        """
        Initialize dataset builder.
        
        Args:
            output_dir: Directory to save tasks
        """
        self.output_dir = Path(output_dir)
        self.output_dir.mkdir(exist_ok=True, parents=True)
        
        self.humaneval_scraper = HumanEvalScraper()
        self.mbpp_scraper = MBPPScraper()
        self.codenet_scraper = CodeNetScraper()
    
    def create_task_directory(
        self,
        task_id: str,
        description: str,
        reference_solution: str,
        test_cases: str,
        metadata: Dict[str, Any]
    ):
        """
        Create a task directory with all required files.
        
        Args:
            task_id: Task identifier
            description: Task description
            reference_solution: Reference solution code
            test_cases: Test cases code
            metadata: Task metadata
        """
        task_dir = self.output_dir / task_id
        task_dir.mkdir(exist_ok=True)
        
        # Write description
        (task_dir / "description.txt").write_text(description, encoding='utf-8')
        
        # Write reference solution
        (task_dir / "reference_solution.py").write_text(reference_solution, encoding='utf-8')
        
        # Write test cases
        (task_dir / "test_cases.py").write_text(test_cases, encoding='utf-8')
        
        # Write metadata
        with open(task_dir / "metadata.json", 'w', encoding='utf-8') as f:
            json.dump(metadata, f, indent=2)
        
        logger.info(f"Created task directory: {task_id}")
    
    def build_dataset(
        self,
        humaneval_count: int = 8,
        mbpp_count: int = 7,
        codenet_count: int = 0
    ):
        """
        Build complete dataset from multiple sources.
        
        Args:
            humaneval_count: Number of HumanEval tasks
            mbpp_count: Number of MBPP tasks
            codenet_count: Number of CodeNet tasks
        """
        logger.info("Building task dataset...")
        
        task_counter = 1
        
        # Fetch HumanEval tasks
        if humaneval_count > 0:
            humaneval_tasks = self.humaneval_scraper.fetch_tasks(humaneval_count)
            for task in humaneval_tasks:
                # TODO: Convert to our format and create directory
                task_counter += 1
        
        # Fetch MBPP tasks
        if mbpp_count > 0:
            mbpp_tasks = self.mbpp_scraper.fetch_tasks(mbpp_count)
            for task in mbpp_tasks:
                # TODO: Convert to our format and create directory
                task_counter += 1
        
        # Fetch CodeNet tasks
        if codenet_count > 0:
            codenet_tasks = self.codenet_scraper.fetch_tasks(codenet_count)
            for task in codenet_tasks:
                # TODO: Convert to our format and create directory
                task_counter += 1
        
        logger.info(f"Dataset build complete! Total tasks: {task_counter - 1}")


def main():
    """Main function to build dataset."""
    builder = TaskDatasetBuilder()
    builder.build_dataset(
        humaneval_count=8,
        mbpp_count=7,
        codenet_count=0
    )


if __name__ == "__main__":
    main()
